# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "kuk.aka@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8876",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/v1/notebook/store": {
            "post": {
                "tags": [
                    "Add"
                ],
                "summary": "Create Note",
                "operationId": "f92662758205ac97205f30554fe06f60",
                "requestBody": {
                    "$ref": "#/components/requestBodies/RegistrationRequest"
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "fio": {
                                            "type": "string",
                                            "example": "Ivanov I.I"
                                        },
                                        "company": {
                                            "type": "string",
                                            "example": "SuperCompany"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "891611111111"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "example@email.ru"
                                        },
                                        "birthdate": {
                                            "type": "string",
                                            "example": "01.01.1990"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "photo.png"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notebook/{id}": {
            "get": {
                "tags": [
                    "Select note"
                ],
                "summary": " Detail note from get",
                "description": "Выборка методом  get",
                "operationId": "2d15187bb08ff257534f59605e84d401",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "fio": {
                                            "type": "string",
                                            "example": "fio"
                                        },
                                        "company": {
                                            "type": "string",
                                            "example": "company"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "phone"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email"
                                        },
                                        "birthdate": {
                                            "type": "string",
                                            "example": "birthdate"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Select note"
                ],
                "summary": " Detail note from post",
                "description": "Выборка метедом post",
                "operationId": "ababe2e95fe318c16de86a9e77a43418",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "fio": {
                                            "type": "string",
                                            "example": "fio"
                                        },
                                        "company": {
                                            "type": "string",
                                            "example": "company"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "phone"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email"
                                        },
                                        "birthdate": {
                                            "type": "string",
                                            "example": "birthdate"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notebook/gets": {
            "get": {
                "tags": [
                    "All notes"
                ],
                "summary": " All notes from get",
                "description": "Get",
                "operationId": "examplesAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "All notes"
                ],
                "summary": " All notes from post",
                "description": "Post",
                "operationId": "examplesAllPost",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/delete/{id}": {
            "delete": {
                "tags": [
                    "Delete"
                ],
                "summary": "Delete Note",
                "operationId": "481d87a13b7516bde50b8a6fe3218881",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "delete todo success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "requestBodies": {
            "RegistrationRequest": {
                "description": "Registration request body",
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "fio",
                                "phone",
                                "email"
                            ],
                            "properties": {
                                "fio": {
                                    "type": "string",
                                    "example": "Ivanov I.I."
                                },
                                "company": {
                                    "type": "string",
                                    "example": "ExampleCo"
                                },
                                "phone": {
                                    "type": "string",
                                    "example": "89991111111"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "example@mail.ru"
                                },
                                "birthdate": {
                                    "type": "string",
                                    "example": "01.01.1990"
                                },
                                "image": {
                                    "type": "string",
                                    "example": "photo.png"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    }
}